+++ loop/transforms/shift_column_transform.py
@@ -1,7 +1,7 @@
 import polars as pl
 
 
-def shift_column(data: pl.DataFrame, shift: int, column: str) -> pl.DataFrame:
+def shift_column_transform(data: pl.DataFrame, shift: int, column: str) -> pl.DataFrame:
 
     '''
     Shift a column by a specified number of periods.
@@ -12,7 +12,7 @@ def shift_column(data: pl.DataFrame, shift: int, column: str) -> pl.DataFrame:
         column (str): Name of column to shift
 
     Returns:
-        pl.DataFrame: DataFrame with shifted target column
+        pl.DataFrame: DataFrame with shifted column
     '''
 
     return data.with_columns(